{"schema":"decodex/commit/1","summary":"Split retrieval merge policy helpers after strict validation.","authority":"manual"}
